use Cursive frontend with command line argument `cli`
Defaults to SDL3, but if run with argument `cli` it will use the Cursive/ncurses frontend. Update timing is slightly different between the two frontends.
