+++ Cargo.toml
@@ -5,8 +5,8 @@ edition = "2024"
 
 [dependencies]
 arrayvec = "0.7.4"
-# cursive = "0.21.1"
-# cursive_core = "0.4.6"
+cursive = "0.21.1"
+cursive_core = "0.4.6"
 image = "0.25.6"
 libc = "0.2.172"
 log = "0.4.27"
